chore(music): share context matching helper
