+++ plugins/plugin-music/src/actions/playbackOp.ts
@@ -22,6 +22,7 @@ import {
 import { classifyPlaybackTransportIntent } from "../utils/playbackTransportIntent";
 import { ProgressiveMessage } from "../utils/progressiveMessage";
 import { resolveMusicGuildIdForPlayback } from "../utils/resolveMusicGuildId";
+import { selectedContextMatches } from "../utils/selectedContextMatches";
 import { mergedOptions, requireMusicConfirmation } from "./confirmation";
 
 function formatFetchProgressDetails(details: unknown): string | undefined {
@@ -47,35 +48,6 @@ function failureResult(
   data?: ActionResultData,
 ): ActionResult {
   return { success: false, text, error, data };
-}
-
-function selectedContextMatches(
-  state: State | undefined,
-  contexts: readonly string[],
-): boolean {
-  const selected = new Set<string>();
-  const collect = (value: unknown) => {
-    if (!Array.isArray(value)) return;
-    for (const item of value) {
-      if (typeof item === "string") selected.add(item);
-    }
-  };
-  collect(
-    (state?.values as Record<string, unknown> | undefined)?.selectedContexts,
-  );
-  collect(
-    (state?.data as Record<string, unknown> | undefined)?.selectedContexts,
-  );
-  const contextObject = (state?.data as Record<string, unknown> | undefined)
-    ?.contextObject as
-    | {
-        trajectoryPrefix?: { selectedContexts?: unknown };
-        metadata?: { selectedContexts?: unknown };
-      }
-    | undefined;
-  collect(contextObject?.trajectoryPrefix?.selectedContexts);
-  collect(contextObject?.metadata?.selectedContexts);
-  return contexts.some((context) => selected.has(context));
 }
 
 function normalizeOp(value: unknown): PlaybackOp | null {
